requirements.txt
pyyaml~=6.0.2
pandas~=2.2.3
openai==1.59.3
numpy~=1.26.4
scikit-learn~=1.5.2
transformers~=4.45.2
chromadb~=0.5.23
langchain~=0.3.14
langchain-chroma~=0.1.4
langchain-community~=0.3.14
langchain-huggingface~=0.1.2
langchain_openai==0.2.14
langchain-text-splitters~=0.3.4
langchain-core~=0.3.29
ollama~=0.4.5
requests~=2.32.3
termcolor==2.5.0
torch~=2.5.1+cu124
sentencepiece~=0.2.0
neo4j~=5.27.0
sentence-transformers~=3.3.1
pillow~=11.0.0
langchain-openai~=0.2.14
torch~=2.5.1+cu124